% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_models.R
\name{copy_models}
\alias{copy_models}
\title{Copy Stan Models}
\usage{
copy_models(pkgname = "staninside", local_location = NULL)
}
\arguments{
\item{pkgname}{a string, the name of the package}

\item{local_location}{a file path representing the desired location
of the local copy. If "default" is provided, the function will use
the user cache directory for the package.}
}
\value{
file path of newly created files
}
\description{
This function extracts the Stan Code from a Package and then
copies it to your user cache locally.
}
\examples{
local_location <- tempdir()
copy_models(pkgname = "staninside", local_location = local_location)
}
