% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_stan_code.R
\name{find_stan_code}
\alias{find_stan_code}
\title{Find Stan Code}
\usage{
find_stan_code(pkgname = "staninside")
}
\arguments{
\item{pkgname}{a string, the name of the package in which to
look for the Stan code.}
}
\value{
a list with the Stan code location within the package
and the text of the code,
}
\description{
This function looks for Stan code within a given R package
a returns a list with the Stan code location within the package
and the text of the code,
}
\examples{
find_stan_code(pkgname = "staninside")
}
