% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_stan_package.R
\name{setup_stan_package}
\alias{setup_stan_package}
\title{Create Package Infrastructure for Stan Package}
\usage{
setup_stan_package(loc = NULL, use_all = TRUE)
}
\arguments{
\item{loc}{a file.path indicating package root with a default of the
current working directory}

\item{use_all}{a logical indicating if each section of Stan code
should have it's own directory with a default of \code{TRUE}}
}
\value{
invisible null
}
\description{
Create directory and helper files for a Stan package using
CmdStanR
}
\examples{

# Create Temporary Directory
temp_dir <- tempdir()

# Now setup the structur
setup_stan_package(loc = temp_dir)

# Examine output

list.files(temp_dir, recursive = TRUE)
}
